python_mcp_host/pyproject.toml
[project]
name = "python-mcp-host"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "ruff>=0.7.0",
    "loguru>=0.7.3",
    "msgpack-python>=0.5.6",
    "numpy>=2.2.6",
    "pyzmq>=27.1.0",
    "google-generativeai>=0.8.0",
]
